Delete uint64 support
The API is a lot more user friendly if you don't have to think about uint64 vs.
int64, especially because an int64 won't survive a round trip if it is positive
(it gets converted into a uint64).

Simplify the API by deleteing uint64 support. We gracefully fail if decoding
can't fit a uint64_t into a int64_t.